:art: add rest begin rest duration
